requirements.txt
Pillow>=10.0.0
easyocr>=1.7.0
customtkinter>=5.2.0
requests>=2.31.0
tenacity>=8.2.0
mypy>=1.7.0
numpy>=1.24.0
pytest>=7.4.0
black>=23.12.0
isort>=5.13.0
opencv-python>=4.8.0 